08-LangChain-Projects/Formula1_Chatbot/pyproject.toml
[tool.poetry]
name = "formula1-chatbot"
version = "0.1.0"
description = ""
authors = ["abdulrehman1004 <rehman.abdul008844@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.14"
langchain-community = "^0.3.14"
streamlit = "^1.41.1"
langchain-google-genai = "^2.0.8"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
astrapy = "^1.5.2"
langchain-qdrant = "^0.2.0"
qdrant-client = "^1.12.2"
langchain-huggingface = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
